RagBackend/requirements.txt
aiofiles==24.1.0
beautifulsoup4==4.13.4
camelot-py==0.11.0
fastapi==0.116.1
google-api-python-client==2.109.0
Jinja2==3.1.6
langchain==0.3.27
langchain_community==0.3.27
langchain_huggingface==0.3.1
langchain_ollama==0.3.6
layoutparser==0.3.4
pdf2image==1.17.0
pdfplumber==0.11.7
Pillow==11.3.0
protobuf==6.31.1
pydantic==2.11.7
PyJWT==2.10.1
PyMySQL==1.1.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
python-docx==1.1.2
Requests==2.32.4
uvicorn==0.35.0
docx2txt==0.9
faiss-cpu>=1.12.0
python-multipart==0.0.6
sentence-transformers>=5.1.0
# Redis async client (Redis Stream task queue + status persistence)
redis[hiredis]>=5.0.0
